Changed commands a little, fine tuned Lua
Commands were changed so that they can be run from the console, but
mini-game specific command will not be able to. New 'reloadlua' command
now. Lua now has a better folder system, and will only load Lua file
directly in the GSMG_MiniGames folder, or directly in a
GSMG_MiniGames/MiniGames/<mini-game name>/ folder. This allow external
Lua files to be used and not run accidentally. Global variable have been
added that point to the Includes and MiniGame folders, to allow
'require's to work better.