Added a Lua library for some needed functions
This is a file I have been working on, and it contains Lua functions to
make general coding easier. The PrintTable function (inspired by Gmod)
will print the entire contents of a table to the console. The wait
function (inspired by ROBLOX) will wait for the amount of seconds. It
should be rarely, if at all, used. It will make the entire server stop,
which is not good, so avoid using it. The table.tostring and
table.fromstring (inspired by ComputerCraft & Gmod) will convert a table
to a full string, and back into a table. This allows for printing
(although, PrintTable does a better job of this) and for saving to a
file.