Added in capability to load java addons
The class path global now works differently, as addon loading has been
added. This means that you can write extra code in java, and have it
automatically be added to the Lua. This is done by having a class in the
path 'gsmg.addon.main' that contains the function 'getClasses'. The
function should return a hashmap containing the name and path of all the
classes that the addon provides, allowing the Lua to have extra classes
in the global table. The 'Vanish'/'gsmg_base_hub' class has been fixed
to a state to allow compiling of the plugin.