Finished Template Mini-Game, Modified events and commands
The template mini-game is now complete, and can be used. I will write an
example mini-game later. The events have been changed so there is now
less code, and a new 'Shutdown' event which is called when the plugin is
disabled or on a reload. The 'reload' command has been implemented. A
new 'chatcolours' class has been added to the Lua, and it contains the
values of all the colours (no functions). README.md updated. The
mini-game loader will now only run files that have the same name as the
directory that it is in.